% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_variant}
\alias{filter_variant}
\title{Title}
\usage{
filter_variant(
  merged_data,
  sample_info,
  edit_rate_indi = 0.05,
  edit_rate_mean = 0.05,
  missing_rate = NA,
  left_n_total = NA,
  left_n_in_each_group = 3
)
}
\arguments{
\item{left_n_in_each_group}{}
}
\description{
Title
}
