% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Title}
\usage{
plot_volcano(
  data,
  FDR_thres = 0.05,
  delta_thres = 0,
  show_up_down = FALSE,
  save = FALSE,
  file = NA
)
}
\arguments{
\item{file}{}
}
\description{
Title
}
