% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_redit}
\alias{read_redit}
\title{read_redit}
\usage{
read_redit(sample_info, soft = "reditools_known")
}
\arguments{
\item{sample_info}{df of 3 cols(filename, sampleid, group)}

\item{soft}{c("reditools_known") currently supporting reditools results.}
}
\value{
matrix of RNA editing sites
}
\description{
read_redit
}
\examples{
read_redit(sample_info, soft="reditools_known")
}
