% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_edit_type_count}
\alias{plot_edit_type_count}
\title{Title}
\usage{
plot_edit_type_count(data = rna_type_count, save = FALSE, file = NA)
}
\arguments{
\item{data}{dataframe, rna_type_count}

\item{save}{whether to save}

\item{file}{FilePath to save}
}
\value{
barplot of rna editing type count
}
\description{
Title
}
\examples{
plot_edit_type_count(rna_type_count)

plot_edit_type_count(rna_type_count, save=TRUE, file="plot_rna_type_count.pdf")
}
